% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGRBasicFunctions.R
\name{bedGraphFromGR}
\alias{bedGraphFromGR}
\title{Convert a GemomicRanges object (1-based close) in a dataframe like bedgraphs (0-based half open)}
\usage{
bedGraphFromGR(gr)
}
\arguments{
\item{gr}{a GenomicRanges object}
}
\value{
a data frame with 4 columns (seqnames, start, end, score)
}
\description{
Convert a GemomicRanges object (1-based close) in a dataframe like bedgraphs (0-based half open)
}
\examples{
gr <- GenomicRanges::GRanges(seqnames = "chr1",
                             ranges = IRanges::IRanges(start = c(1, 11),
                                                       end = c(10, 12)),
                             score = c(20, 30))
bedGraphFromGR(gr)
#   seqnames start end score
# 1     chr1     0  10    20
# 2     chr1    10  12    30
}
