% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGRoverlapFunctions.R
\name{findOverlapAsClusters}
\alias{findOverlapAsClusters}
\title{Get a dataframe summarizing the clusters created by overlaps
from a GRangesList where each element is only in one cluster}
\usage{
findOverlapAsClusters(myGRsToOverlap)
}
\arguments{
\item{myGRsToOverlap}{a GRangesList with items to overlap}
}
\value{
A dataframe with one column per item in the list with colnames identicals to names of the input
each row is a cluster of overlapping items
each value is a vector with the indices of the items in the cluster.
}
\description{
Get a dataframe summarizing the clusters created by overlaps
from a GRangesList where each element is only in one cluster
}
\examples{
gr1 <- GenomicRanges::GRanges(seqnames = "chr1",
ranges = IRanges::IRanges(start = c(1, 11, 199),
                         end = c(10, 12, 200)),
score = c(20, 30, 100))
gr2 <- GenomicRanges::GRanges(seqnames = "chr1",
                             ranges = IRanges::IRanges(start = c(5, 101),
                                                       end = c(15, 102)),
                             score = c(1, 2))
gr3 <- GenomicRanges::GRanges(seqnames = "chr1",
                             ranges = IRanges::IRanges(start = c(1, 100),
                                                       end = c(4, 120)),
                             score = c(10, 20))
grL <- list(first = gr1, second = gr2, third = gr3)
findOverlapAsClusters(grL)
}
