% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRTrackLayerBasicFunctions.R
\name{getTSSinUCSCFormatFromEnsemblGTF}
\alias{getTSSinUCSCFormatFromEnsemblGTF}
\title{Get a GRange with TSS in UCSC format from a path to a gtf file in Ensembl format}
\usage{
getTSSinUCSCFormatFromEnsemblGTF(myGTF)
}
\arguments{
\item{myGTF}{a file path to a gtf file with ensembl chromosome annotation (no "chr")}
}
\value{
a Genomic Range with all TSS (entries might be duplicated) in UCSC format (with "chr")
}
\description{
Get a GRange with TSS in UCSC format from a path to a gtf file in Ensembl format
}
\examples{
\dontrun{
tss <- getTSSinUCSCFormatFromEnsemblGTF("Homo_sapiens.GRCh38.95.gtf.gz")
}
tss <- getTSSinUCSCFormatFromEnsemblGTF(paste0(system.file("tests", package="usefulLDfunctionsGR"),
                                               "/Homo_sapiens.GRCh38.95_491firstLines.gtf.gz"))
}
