% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGRBasicFunctions.R
\name{grFromBedFile}
\alias{grFromBedFile}
\title{create a GenomicRanges object from a BED}
\usage{
grFromBedFile(fn)
}
\arguments{
\item{fn}{a path to the BED file (it can be gzipped) (0-based half open)}
}
\value{
a Genomic ranges with intervals (1-based closed)
}
\description{
create a GenomicRanges object from a BED
}
\examples{
tests_dir <- system.file("tests", package="usefulLDfunctionsGR")
test_bed <- file.path(tests_dir, "test6colWithHeader.bed")

# Load the bed with no header in a dataframe
grBED <- grFromBedFile(test_bed)
}
