% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGRBasicFunctions.R
\name{grSortedSimplifiedFromNarrowPeak}
\alias{grSortedSimplifiedFromNarrowPeak}
\title{create a GenomicRanges object from a narrowPeak with only one entry per peak (even if multiple summits) sorted by score}
\usage{
grSortedSimplifiedFromNarrowPeak(fn)
}
\arguments{
\item{fn}{a path to the narrowPeak file (it can be gzipped)}
}
\value{
a Genomic ranges sorted by score decreasing with one line per peak, the summit kept is the one with the higher score
}
\description{
create a GenomicRanges object from a narrowPeak with only one entry per peak (even if multiple summits) sorted by score
}
