% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGRBasicFunctions.R
\name{narrowPeakDFFromGR}
\alias{narrowPeakDFFromGR}
\title{create a data.frame from a GenomicRanges object from a narrowPeak}
\usage{
narrowPeakDFFromGR(gr)
}
\arguments{
\item{gr}{a GenomicRanges which contains all information for writting a narrowPeak}
}
\value{
a data.frame with intervalls 0-based half open, ready to be written in a file.
}
\description{
create a data.frame from a GenomicRanges object from a narrowPeak
}
